POC/Filtering/server/pyproject.toml
[tool.poetry]
name = "dddbase"
version = "0.1.0"
description = ""
authors = ["Harry"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.86.0"
uvicorn = "^0.19.0"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
fastapi-jwt-auth = "^0.5.0"
elasticsearch = "7.14.0"
elasticsearch-dsl = "^7.4.0"
motor = "^3.1.1"
geopy = "^2.2.0"
python-dateutil = "^2.8.2"
pyfunctional = "^1.4.3"

[tool.poetry.dev-dependencies]
uvicorn = "^0.19.0"
pydantic = "^1.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
